requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
sqlalchemy>=2.0.20
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.3.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
Pillow>=10.0.0
